+++ src/wsgi.py
@@ -1,0 +1,3 @@
+from ansible_link import main
+
+application = main()